backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
sqlalchemy
psycopg[binary]
python-dotenv
httpx
pydantic
pydantic-settings
apscheduler
geopy
python-dateutil
alembic